requirements.txt
amqp==5.1.1
asgiref==3.7.2
asn1crypto==1.5.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.34.0
colorama==0.4.6
Django==4.2.4
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
environs==9.5.0
idna==3.4
lxml==4.9.3
marshmallow==3.20.1
neapolitan==23.11
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.7
pyaxmlparser==0.3.28
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==6.4.0
text-unidecode==1.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
